make Proof a final class
